Remove warnings from tests (#1395)

+++ test/models/test_jastrow.py
@@ -24,7 +24,7 @@ def test_Jastrow(dtype):
     hi = nk.hilbert.Spin(1 / 2, N)
     g = nk.graph.Chain(N)
 
-    ma = nk.models.Jastrow(dtype=dtype)
+    ma = nk.models.Jastrow(param_dtype=dtype)
     _ = ma.init(nk.jax.PRNGKey(), hi.random_state(nk.jax.PRNGKey()))
 
     vmc = nk.VMC(
